Removed old, unused batch calculation scripts. Renamed the current version to batch_calculate_e2e_perf.py
